Days 2 and 3
Day 2 part 1 is a bit of a mess, but part 2 is pretty elegant: For each character position in the ids, remove that character from all elements, sort, and look for adjacent identical strings (which are also are what you need to return). It's M * N log N where N is the number of lines and M the number of characters per string; can probably do better.

Day 3 was timed so it's a bit messy; took just under 10 minutes.